bugfix: some files were installed as user 502
A few problems contributed:

1) A few files were specified to install as my user in the installer
pmdoc xml files.
2) OS X PackageMaker is inept. It ignores the specifications from the
installer pmdoc xml files, and packages the files to install as the user
who is building the package.

Problem 1 was solved by modifying the pmdoc xml files.  To work around
problem 2, I resorted running PackageMaker as root, which appears to be
satisfactory.

A test is added to ensure future releases don't go out with user-owned
files.