Change the use of ARCH to HOSTTYPE which exists by default across 10.6 and 10.7
I'm ultimately not sure where the use of the ARCH variable came from, because it didn't exist on my systems. Looking around an Intel Mac mini still on 10.6 and the Intel iMac on 10.7 I'm currently using, I found that the HOSTTYPE variable is set appropriate in the environment to either i386 or x86_64. I initially tried to use the arch command's output, but that appears to return i386 no matter what, otherwise ppc if that is the processor architecture in use. Pending some testing, this change appears to work well for any architecture, on Snow Lepord, or Lion, and has the chance to work very well outside of those two OSes.