Add a GitHub workflow to build installers
It is really convenient to be able to use GitHub Actions to run
automated tasks, such as creating the Git for macOS installer.

This adds a GitHub workflow definition to do exactly that.

It can be triggered manually from the Actions tab of the
`git_osx_installer` repository on GitHub. For details, see
https://github.blog/changelog/2020-07-06-github-actions-manual-triggers-with-workflow_dispatch

Alternatively, the `push:` line can be uncommented so that a simple push
triggers the workflow.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>