uninstall.sh: correctly interpret "yes"
The uninstall script takes an optional "--yes" argument to avoid
interactively asking the user if it is OK to uninstall. However,
this causes problems when the user wants to upgrade because the
condition is actually the opposite!

This is a problem for "brew upgrade --cask git" because the first
uninstall fails due to needing sudo access, and the second one that
calls with sudo uses --yes.

Signed-off-by: Derrick Stolee <dstolee@microsoft.com>